+++ forge/changelog.md
@@ -1,2 +1,4 @@
 - Fixed game freezing when far away from the world origin
 - Fixed lag spikes
+- Reworked reflectivity config
+- Reworked occlusion config